[refactoring][01] geomapshark -> geocity
