+++ .gitignore
@@ -5,15 +5,15 @@ __pycache__/
 /print/*.conf
 /print/*.qgs~
 /static/images/print/STE_Signature.png
-geomapshark/static/
+geocity/static/
 gunicorn_log.log
 static/tempfiles/
-geomapshark/protected-static/
+geocity/protected-static/
 env.yaml
 /docker-compose.override.yml
 /permit_requests_uploads
 /administrative_entity_customization
-geomapshark/permit_requests_uploads/
+geocity/permit_requests_uploads/
 private_documents/
 /printed_permits
 *.env
