+++ README.md
@@ -60,25 +60,25 @@ CLEAR_PUBLIC_SCHEMA_ON_FIXTURIZE=true
 
 Run tests in a the running container
 ```bash
-docker-compose exec web python manage.py test --settings=geomapshark.settings_test
+docker-compose exec web python manage.py test --settings=geocity.settings_test
 ```
 
 Run a specific test in the running container (adding the `--keepdb` flag speeds up iterations)
 ```bash
-docker-compose exec web python manage.py test --settings=geomapshark.settings_test --keepdb permits.tests.test_a_permit_request.PermitRequestTestCase.test_administrative_entity_is_filtered_by_tag
+docker-compose exec web python manage.py test --settings=geocity.settings_test --keepdb permits.tests.test_a_permit_request.PermitRequestTestCase.test_administrative_entity_is_filtered_by_tag
 ```
 
 Test for report generation will fail when run in the running web container, because they spawn a test server to allow other container to communicate with it. You can run them in their own container, but need first to stop the running web container.
 
 ```bash
 docker-compose stop web
-docker-compose run --service-ports --name=web --rm --entrypoint="" web python manage.py test --settings=geomapshark.settings_test --keepdb reports
+docker-compose run --service-ports --name=web --rm --entrypoint="" web python manage.py test --settings=geocity.settings_test --keepdb reports
 ```
 
 These tests compare generated PDFs to expected PDFs. To regenerated the expected images, you need to provide the following env var to the docker-compose run command `-e TEST_UPDATED_EXPECTED_IMAGES=TRUE`. By definition, when doing so, tests will succeed, so don't forget to manually review the changes to the files.
 
 ```bash
-docker-compose run --service-ports --name=web --rm --entrypoint="" -e TEST_UPDATED_EXPECTED_IMAGES=TRUE web python manage.py test --settings=geomapshark.settings_test --keepdb reports
+docker-compose run --service-ports --name=web --rm --entrypoint="" -e TEST_UPDATED_EXPECTED_IMAGES=TRUE web python manage.py test --settings=geocity.settings_test --keepdb reports
 ```
 
 ### Linting
