+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
             args:
                 dev_dependencies: "${DEV_DEPENDENCIES}"
         entrypoint: /code/entrypoint.sh
-        command: "gunicorn geomapshark.wsgi -b :9000 --error-logfile gunicorn_log.log --workers=2 --threads=4 --worker-class=gthread"
+        command: "gunicorn geocity.wsgi -b :9000 --error-logfile gunicorn_log.log --workers=2 --threads=4 --worker-class=gthread"
         volumes:
             - ${PRIVATE_DOCUMENTS_DIR}:/private_documents
             - ${ARCHIVE_DIR}:/archive
