+++ entrypoint_demo.sh
@@ -17,7 +17,7 @@ cd /code
 python3 manage.py migrate
 # django-constance models
 python3 manage.py migrate database
-mkdir -p /code/geomapshark/static/
+mkdir -p /code/geocity/static/
 echo yes | python3 manage.py compilemessages -l fr
 echo yes | python3 manage.py collectstatic
