+++ env.demo
@@ -34,7 +34,7 @@ EMAIL_TO_CONSOLE=false
 DEFAULT_FROM_EMAIL=your_noreply_email
 GEOPORTAL_URL=your_geoportal_base_url
 PAGINATE_BY=20 # For list display
-STATIC_FILES_ABSOLUTE_PATH=/code/geomapshark/static
+STATIC_FILES_ABSOLUTE_PATH=/code/geocity/static
 MIN_START_DELAY=2 # Min number of days before a permit can be delivered
 OL_MAP_HEIGHT=500
 WMTS_GETCAP=https://wmts.asit-asso.ch/wmts?request=GetCapabilities
