+++ geocity/settings_dev.py
@@ -19,9 +19,7 @@ MIDDLEWARE = [
 
 INTERNAL_IPS = ["127.0.0.1"]
 
-DEBUG_TOOLBAR_CONFIG = dict(
-    SHOW_TOOLBAR_CALLBACK="geomapshark.settings_dev.show_toolbar"
-)
+DEBUG_TOOLBAR_CONFIG = dict(SHOW_TOOLBAR_CALLBACK="geocity.settings_dev.show_toolbar")
 
 SITE_HTTPS = False
 SESSION_COOKIE_HTTPONLY = False
