+++ geocity/views.py
